% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-text.R
\name{add_text}
\alias{add_text}
\title{Obtain Text from Download Links and Add to Data Frame}
\usage{
add_text(df, all = FALSE)
}
\arguments{
\item{df}{data frame produced by \code{\link{get_all_comments}} or \code{\link{get_all_documents}}}

\item{all}{logical; TRUE if you want to include all attachments for each comment,
FALSE if you want to include only one. The default value is FALSE. If you choose
TRUE, note that the attachments may include duplicated text, so this should be considered
for further analysis.}
}
\value{
data frame with all columns of \code{df} as well as an additional column \code{text} that
contains the text from the attachment in the download link for that commend ID (if present).
}
\description{
Obtain Text from Download Links and Add to Data Frame
}
\examples{
\dontrun{
result <- get_all_documents(documentId = c(
  "FDA-2012-S-1144-0322",
  "NHTSA-2008-0060-0827", "DOT-OST-2018-0206-0008", "CMS-2018-0104-0001"
))
resul <- add_text(result)
}
}
