% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_auth}
\alias{check_auth}
\title{Verify Key is Available}
\usage{
check_auth(key)
}
\arguments{
\item{key}{the API key passed in the function call; this may be NULL if the user has
chosen to set up the key as an environmental variable instead with the function
\code{\link{set_datagov_key}}}
}
\description{
This function checks if the user has the API key set up or passed as an argument
}
\keyword{internal}
