% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-all-comments.R
\name{check_continue}
\alias{check_continue}
\title{Check if the User Wants to Continue}
\usage{
check_continue(num_elements)
}
\arguments{
\item{num_elements}{of elements associated with the url}
}
\description{
In an interactive session, check if the user wants to continue when
the number of requests means the call will take an extended amount of
time (over 500 requests means the call will take over an hour due to
the 500 per hour request limit).
}
