% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-params.R
\name{check_date}
\alias{check_date}
\title{Helper function to check validity of date parameters}
\usage{
check_date(dates, param)
}
\arguments{
\item{dates}{date vector or single string to be checked}

\item{param}{name of the parameter}
}
\value{
error message; empty string if no issues found
}
\description{
Helper function to check validity of date parameters
}
