% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-params-comments.R
\name{check_non_id}
\alias{check_non_id}
\title{Check if Invalid Arguments Are Provided When ID Argument is Provided}
\usage{
check_non_id(arg_list, id)
}
\arguments{
\item{arg_list}{named list containing arguments to the one of the \verb{*_construct_url} functions.}

\item{id}{which type of ID to check ("docketId", "documentId", or "commentId")}
}
\description{
Because the structure of the url is different if the user provides an id,
more specifically a "docketId", "documentId", or "commentId", this function
checks if the user has supplied additional arguments that can't be included into
the valid API call, and constructs an error accordingly.
}
\keyword{internal}
