% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-text.R
\name{choose_element}
\alias{choose_element}
\title{Choose Link Based on Format}
\usage{
choose_element(links)
}
\arguments{
\item{links}{character vector of links}
}
\description{
Choose Link Based on Format
}
\keyword{internal}
