% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-document-url.R
\name{collapse_values}
\alias{collapse_values}
\title{Collapse Argument Values}
\usage{
collapse_values(x, arguments)
}
\arguments{
\item{x}{name of element in the named list arguments}

\item{arguments}{named list containing arguments passed to the
\code{\link{construct_document_url}} function}
}
\description{
Collapse values to a comma separated string when a list
argument contains multiple values that need to be included in
the url.
}
