% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-docket-url.R
\name{construct_docket_url}
\alias{construct_docket_url}
\title{Create a URL for the Dockets Endpoint}
\usage{
construct_docket_url(
  key = NULL,
  docketId = NULL,
  agencyId = NULL,
  docketType = NULL,
  searchTerm = NULL,
  lastModifiedDate = NULL,
  sort = NULL,
  page_number = 1,
  page_size = 250
)
}
\arguments{
\item{key}{the API key passed in the function call; this may be NULL if the user has
chosen to set up the key as an environmental variable instead with the function
\code{\link{set_datagov_key}}. You can use "DEMO_KEY" for a very limited number of calls if needed.
with the \code{docketId} argument.}

\item{docketId}{character string representing one valid docket ID, for example, "CMS-2014-0063".}

\item{agencyId}{character string representing a government agency, for example, "CMS" or "EPA" or a character
vector of multiple agencies}

\item{docketType}{the type of docket, valid types are "Rulemaking" and "Nonrulemaking"}

\item{searchTerm}{character string representing a search term to filter the results for or a character
vector with multiple search terms to filter for.}

\item{lastModifiedDate}{a character vector of length 1, representing a single posted date to filter for,
or a vector of length 2, representing a date range beginning with the first element and ending at the second.
Each element must be in the format yyyy-MM-dd HH:mm:ss.}

\item{sort}{character string representing which value you want to sort the results by, supported variables are
"title", "docketId", "lastModifiedDate"}

\item{page_number}{character string representing which page to retrieve the data; useful when
output is on multiple pages. Valid values are between 1 and 20. Default value is 1.}

\item{page_size}{character string representing how many elements should be on each page. Valid values
are between 5 and 250. The default value is 250.}
}
\description{
Takes the parameters available for the documents endpoint of the regulations.gov API, as
described \href{https://open.gsa.gov/api/regulationsgov/#searching-for-documents}{here} under the
documents section under API calls, and constructs a URL that can be used to retrieve the
data specified. To understand how these parameters relate to the data, it may help to go the
the regulations.gov search function and go to a \href{https://www.regulations.gov/docket/CMS-2014-0063}{specific docket}
or \href{https://www.regulations.gov/document/CMS-2014-0063-0001}{document} and look at the details section.
}
\examples{
url <- construct_docket_url(agencyId = c("CMS", "EPA"),
lastModifiedDate =  c("2020-01-02 12:00:00", "2020-02-02 12:00:00" ), key = "DEMO_KEY")
url <- construct_docket_url(docketId = "CMS-2014-0063-0001", key = "DEMO_KEY")
}
