% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{convert_df}
\alias{convert_df}
\title{Helper Function to Convert get_data output to data frame}
\usage{
convert_df(parsed_data)
}
\arguments{
\item{parsed_data}{output from \code{jsonlite::fromJSON(httr::content(resp, "text"),simplifyVector = TRUE,flatten = TRUE)}
in the get_data function.}
}
\value{
a data frame of 1 row, where each column represents
data a field of information from the nested list.
}
\description{
This function is for use with the \code{df=TRUE} argument with
\code{\link{get_data}}. It converts the nested list containing
the data into a data frame.
}
