% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{convert_time}
\alias{convert_time}
\title{Convert Date of the form '2020-07-22T18:28:40Z' to Eastern Time}
\usage{
convert_time(date)
}
\arguments{
\item{date}{of the form '2020-07-22T18:27:52Z'}
}
\value{
date of the form '2020-07-22 14:28:40'
}
\description{
Convert Date of the form '2020-07-22T18:28:40Z' to Eastern Time
}
\keyword{internal}
