% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{extract_comment_meta}
\alias{extract_comment_meta}
\title{Extract Metadata for Specific Comments}
\usage{
extract_comment_meta(comment_link, key = NULL)
}
\arguments{
\item{comment_link}{a link of the form https://api.regulations.gov/v4/comments/COMMENT_ID_HERE for which
detailed data will be obtained.}

\item{key}{the API key passed in the function call; this may be NULL if the user has
chosen to set up the key as an environmental variable instead with the function}
}
\value{
a data frame containing the metadata for the comment ID, including, if present
the attachment links for that comment.
}
\description{
Extract Metadata for Specific Comments
}
\examples{
\dontrun{
extract_comment_meta("https://api.regulations.gov/v4/comments/COMMENT_ID_HERE")
}
}
\keyword{internal}
\keyword{utils::globalVariables(".")}
