% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-meta.R
\name{extract_meta}
\alias{extract_meta}
\title{Extract Metadata for Specific Documents or Comments}
\usage{
extract_meta(link, key = NULL)
}
\arguments{
\item{link}{a link of the form https://api.regulations.gov/v4/comments/ID_HERE for
the comment or document ID for which detailed data will be obtained.}

\item{key}{the API key passed in the function call; this may be NULL if the user has
chosen to set up the key as an environmental variable instead with the function
\code{\link{set_datagov_key}}. You can use "DEMO_KEY" for a very limited number of calls if needed.}
}
\value{
a data frame containing the metadata for the comment or document ID, including, if present
the attachment links.
}
\description{
Extract Metadata for Specific Documents or Comments
}
\examples{
\dontrun{
# get comment metadata for a specific comment
extract_meta("https://api.regulations.gov/v4/comments/COMMENT_ID_HERE")
# get comment metadata for a specific document
extract_meta("https://api.regulations.gov/v4/documents/DOCUMENT_ID_HERE")
}
}
\keyword{internal}
\keyword{utils::globalVariables(".")}
