% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-text.R
\name{extract_text}
\alias{extract_text}
\title{Extract Text from Given Link}
\usage{
extract_text(link)
}
\arguments{
\item{link}{a single download link for a document or comment. For example,
"https://downloads.regulations.gov/FDA-2012-S-1144-0322/attachment_1.pdf".}
}
\value{
text extracted from the given download link
}
\description{
Extract Text from Given Link
}
\keyword{internal}
