% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{find_element}
\alias{find_element}
\title{Find Element in a Nested List with the Given Name}
\usage{
find_element(nested_list, search_for)
}
\arguments{
\item{nested_list}{a nested list, such as the type returned by fromJSON}

\item{search_for}{a character string that is the name of the element we want to find in \code{nested_list}}
}
\description{
Find the value of the element with the given name in
a deeply nested list. Based off of advice
\href{https://stackoverflow.com/questions/58400176/r-find-object-by-name-in-deeply-nested-list}{here}.
}
