% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-all-documents.R
\name{get_all_documents}
\alias{get_all_documents}
\title{Get Document Metadata for Given Documents or Docket}
\usage{
get_all_documents(endpoint = "document", ..., quiet = TRUE, key = NULL)
}
\arguments{
\item{endpoint}{API endpoint that you want to use for the API call.
Accepted options are "document" or "docket".}

\item{...}{arguments passed to \code{\link[=construct_document_url]{construct_document_url()}} if
\code{endpoint = "document"} (the default) or \code{\link[=construct_docket_url]{construct_docket_url()}} if
\code{endpoint = "docket"}}

\item{quiet}{logical; \code{FALSE} if you want detailed information about the URLs and
number of elements to be printed to the console. Default value is \code{TRUE}}

\item{key}{the API key passed in the function call; this may be NULL if the user has
chosen to set up the key as an environmental variable instead with the function
\code{\link{set_datagov_key}}. You can use "DEMO_KEY" for a very limited number of calls if needed.}
}
\value{
a data frame
}
\description{
Get Document Metadata for Given Documents or Docket
}
\examples{
\dontrun{
# get all document data for documents associated with docket FDA-2009-N-0501
result <- get_all_documents(docketId ="FDA-2009-N-0501")
# get all document data for document IDs
result <- get_all_documents(documentId = c("FDA-2012-S-1144-0322",
 "NHTSA-2008-0060-0827", "DOT-OST-2018-0206-0008", "CMS-2018-0104-0001"))
}
}
