% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-all-comments.R
\name{get_comment_links}
\alias{get_comment_links}
\title{Get Comment Links}
\usage{
get_comment_links(docs, key, quiet = TRUE)
}
\arguments{
\item{docs}{a nested list containing data from a valid API call
containing documents and their corresponding objectIds}

\item{key}{valid API key}

\item{quiet}{logical; TRUE if you want messages printed,
FALSE otherwise}
}
\description{
Extract objectIds from nested list, collect comments corresponding
to each \code{objectId}, and return the links of comment links obtained so
detailed comment-specific metadata can be obtained.
}
\keyword{internal}
