% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Acquire Data for the Given Url}
\usage{
get_data(url, df = FALSE, quiet = TRUE, ...)
}
\arguments{
\item{url}{a url containing a valid API call}

\item{df}{logical, \code{TRUE} if you want the output as a data frame.
Default value is \code{FALSE}}

\item{quiet}{logical with default \code{TRUE}, passed to \code{\link[httr:RETRY]{httr::RETRY()}}.
If \code{TRUE} this function does not print messages when \code{\link[httr:RETRY]{httr::RETRY()}} has
to retry getting a request due to hitting a rate limit.}

\item{...}{additional arguments to pass to \code{\link[httr:RETRY]{httr::RETRY()}}}
}
\description{
Acquire Data for the Given Url
}
\examples{
\dontrun{
get_data("https://api.regulations.gov/v4/documents?filter[docketId]=CMS-2014-0063&api_key=DEMO_KEY")
}
}
