% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data_by_page}
\alias{get_data_by_page}
\title{Acquire Data for the Given Page and Url}
\usage{
get_data_by_page(url, page_number = 1)
}
\arguments{
\item{url}{a url containing an API call}

\item{page_number}{page number to obtain data for, default value is 1}
}
\description{
Acquire Data for the Given Page and Url
}
\examples{
\dontrun{
get_data_by_page("https://api.regulations.gov/v4/documents?
filter[docketId]=CMS-2014-0063&page[size]=250&page[number]=1&api_key=DEMO_KEY",
page_number = 2)
}
}
