% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_df}
\alias{get_df}
\title{Convert Element of Nested List from jsonlite::fromJSON into Data Frame}
\usage{
get_df(element)
}
\arguments{
\item{element}{a named nested list}
}
\value{
the element converted to a data frame
}
\description{
Can be used iteratively in conjunction with \code{bind_rows} to convert the
entire nested list into a data frame.
}
