% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-text.R
\name{get_text}
\alias{get_text}
\title{Get Text for Given Link}
\usage{
get_text(link, all)
}
\arguments{
\item{link}{from fileUrl column; may be multiple links comma-separated.}

\item{all}{logical; TRUE if you want to include all attachments for each comment,
FALSE if you want to include only one. The default value is FALSE. If you choose
TRUE, note that the attachments may include duplicated text, so this should be considered
for further analysis.}
}
\description{
Get Text for Given Link
}
\keyword{internal}
