% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{iterate_over_pages}
\alias{iterate_over_pages}
\title{Get All the Elements When Iteration Over Pages May Be Necessary}
\usage{
iterate_over_pages(url, quiet = TRUE)
}
\arguments{
\item{url}{a valid API url that includes the \code{page[size]} and \code{page[number]}
parameters.}

\item{quiet}{logical; FALSE if you want detailed information printed to the console,
TRUE otherwise. FALSE is recommended for longer calls to monitor the progress.}
}
\value{
a nested list containing all elements corresponding to the given url.
Functionality is not yet implemented for obtaining all elements when there are
more than 5000 elements.
}
\description{
Get all elements corresponding to a given url, since the API allows
at most 20 pages, \code{page[number]} parameter of the url, each of
which can contain at most 250 elements, the \code{page[size]} parameter.
If there are less than 250 elements, then no iteration is needed because
all elements are on the first page. Otherwise, multiple API calls will
be needed to obtain all the results.
}
\examples{
\dontrun{iterate_over_pages("https://api.regulations.gov/v4/comments?filter
[commentOnId]=09000064816e1a41&page[size]=250&page[number]=1&
sort=lastModifiedDate,documentId&api_key=DEMO_KEY")
}
}
