% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulationsgov-package.R
\docType{package}
\name{regulationsgov-package}
\alias{regulationsgov}
\alias{regulationsgov-package}
\title{regulationsgov: Accessing Regulations.gov Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package facilitates accessing data through the Regulations.gov API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/q-w-a/regulationsgov}
  \item Report bugs at \url{https://github.com/q-w-a/regulationsgov/issues}
}

}
\author{
\strong{Maintainer}: Quinn White \email{qwhite@smith.edu} (\href{https://orcid.org/0000-0001-5399-0237}{ORCID})

}
\keyword{internal}
