% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set_datagov_key}
\alias{set_datagov_key}
\title{Set Data.gov Key as an Environmental Variable}
\usage{
set_datagov_key(key)
}
\arguments{
\item{key}{valid Data.gov API key obtained \href{https://api.data.gov/signup/}{here}}
}
\description{
This function provides a way to store the Data.gov API key as an
environmental variable for convenience when working with this package.
If you run this function, you will not have to supply the API key
every time you use a function that retrieves data from Regulations.gov.
Otherwise, you will need to supply the API key when needed through
the \code{key} argument.
}
\examples{
\dontrun{
set_datagov_key("PUT_KEY_HERE")
}
}
