% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{skip_if_no_key}
\alias{skip_if_no_key}
\title{Function for Skipping Tests if Authorization is Unavailable}
\usage{
skip_if_no_key()
}
\description{
Function for Skipping Tests if Authorization is Unavailable
}
