% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-params-dockets.R
\name{validate_params_dockets}
\alias{validate_params_dockets}
\title{Validate the Parameters Passed to the construct_docket_url Function}
\usage{
validate_params_dockets(arg_list)
}
\arguments{
\item{arg_list}{named list containing arguments to the \code{\link[=construct_docket_url]{construct_docket_url()}} function.}
}
\description{
Ensure the arguments the \code{construct_document_url} function receives
aren't invalid. This helps to prevent making failed API calls.
}
\keyword{internal}
