% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_element}
\alias{find_element}
\title{Find Element in a Nested List with the Given Name}
\usage{
find_element(nested_list, search_for)
}
\arguments{
\item{nested_list}{a nested list, such as the type returned by fromJSON}

\item{search_for}{a character string that is the name of the element we want to find in \code{nested_list}}
}
\value{
any values of the list where the name contains the substring \code{search_for}
}
\description{
Find the value of the element that contains the given string in its name in
a nested list. Based off of advice
\href{https://stackoverflow.com/questions/58400176/r-find-object-by-name-in-deeply-nested-list}{here}.
}
\examples{
{
# create list of data frames for example
df_nest <- list(Y = data.frame(x.test = c(1,2,3)),
Z = data.frame(w = c("a","b","c")))

# extract element containing 'test'
find_element(df_nest, "test")

# create nested list for example
list_nest <- list(Y = c(1,2,3),
z = list(x.test = c(1,2,3),
y = c("no", "yes")))
find_element(list_nest, "test")
}
}
