% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_all_comments}
\alias{get_all_comments}
\title{Get All Comments for a Specific Docket or Document}
\usage{
get_all_comments(docketId = NULL, documentId = NULL, test = FALSE)
}
\arguments{
\item{docketId}{character string; the docket ID for the comments needed. If you're providing a document
ID with the \code{document_id} argument, this can be left \code{NULL}.}

\item{documentId}{character string; the document ID for the comments needed.}

\item{test}{logical; to test if the output is as expected without using up many API calls.}
}
\value{
a data frame containing the comment metadata, including the download links
}
\description{
This function automates the process of obtaining the metadata,
including download links for attachments for
all comments corresponding to a specific document
or docket ID as multiple API calls will be necessary.
}
\examples{
\dontrun{
# retrieve all comments for docket CMS-2014-0063
comments_CMS_2014_0063 <- get_all_comments(docketId = "CMS-2014-0063")
}
}
