% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Acquire Data for the Given Url}
\usage{
get_data(url, df = FALSE)
}
\arguments{
\item{url}{a url containing a valid API call}

\item{df}{logical, \code{TRUE} if you want the output as a data frame.
Default value is \code{FALSE}}
}
\description{
Acquire Data for the Given Url
}
\examples{
\dontrun{
get_data("https://api.regulations.gov/v4/documents?filter[docketId]=CMS-2014-0063&api_key=DEMO_KEY")
}
}
