% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data_by_page}
\alias{get_data_by_page}
\title{Acquire Data for the Given Page and Url}
\usage{
get_data_by_page(url, page_number = 1, quiet = TRUE)
}
\arguments{
\item{url}{a url containing an API call}

\item{page_number}{page number to obtain data for, default value is 1}

\item{quiet}{logical; FALSE if you want detailed information printed to the console,
TRUE otherwise. FALSE is recommended for longer calls to monitor the progress.}
}
\description{
Get data for the given page number. Used for iterating over pages when
this is necessary to obtain all the data.
}
\examples{
\dontrun{
get_data_by_page("https://api.regulations.gov/v4/documents?
filter[docketId]=CMS-2014-0063&page[size]=250&page[number]=1&api_key=DEMO_KEY",
page_number = 2)
}
}
