% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{iterate_over_pages}
\alias{iterate_over_pages}
\title{Get All the Elements When Iteration Over Pages May Be Necessary}
\usage{
iterate_over_pages(url)
}
\arguments{
\item{url}{a valid API url that includes the \code{page[size]} and \code{page[number]}
parameters.}
}
\description{
Get all elements corresponding to a given url, since the API allows
at most 20 pages, \code{page[number]} parameter of the url, each of
which can contain at most 250 elements, the \code{page[size]} parameter.
If there are less than 250 elements, then no iteration is needed because
all elements are on the first page. Otherwise, multiple API calls will
be needed to obtain all the results.
}
\examples{
\dontrun{iterate_over_pages("https://api.regulations.gov/v4/comments?filter
[commentOnId]=09000064816e1a41&page[size]=250&page[number]=1&
sort=lastModifiedDate,documentId&api_key=DEMO_KEY")
}
}
